+++ 30_sources/CHANGES.md
@@ -7,6 +7,7 @@
 ## enhancement
 
 * now support unlimited extracts #268
+* improve configuration of legend and lyrics #270
 
 ## Compatibility